refactor(wire/user): move UserClass to user/project submodule
